Revert "refactor(sensor_configuration): extracted profile widgets into their own file"
This reverts commit 325263a94454a31d5b02b136f0382e53ed9381ea.
